User input and training function
