--- c2c-translator/data/dataset/simpler_code_examples/cpp/while.cpp
@@ -1,10 +1,0 @@
-#include <iostream>
-
-int main(){
-	int count = 0;
-    while (count < 5) {
-        std::cout<<count<<std::endl;
-        count += 1;
-    }
-	return 0;
-}
