--- c2c-translator/data/dataset/simpler_code_examples/java/while.java
@@ -1,9 +1,0 @@
-public class Main {
-    public static void main(String[] args) {
-      int count = 0;
-      while (count < 5) {
-        System.out.println(count);
-        count += 1;
-      }  
-    }
-  }
