--- c2c-translator/data/dataset/simpler_code_examples/python/while.py
@@ -1,4 +1,0 @@
-count = 0
-while count < 5:
-    print(count)
-    count += 1
