+++ c2c-translator/data/parallel_corpus/sum_two_num.cpp
