+++ c2c-translator/data/parallel_corpus/while.cpp
@@ -1,5 +1,10 @@
-int j = 0;
-while (j < 5) {
-    std::cout<<j;
-    j = j + 1;
+#include <iostream>
+
+int main(){
+	int count = 0;
+    while (count < 5) {
+        std::cout<<count<<std::endl;
+        count += 1;
+    }
+	return 0;
 }
