+++ c2c-translator/data/translations/translation_blocks_from_python.cpp
@@ -1,0 +1,21 @@
+while (j < 5) {
+    std::cout<<j;
+    if (j==2) {
+        std::cout<<"j has the value 2";
+        int a = 5;
+    }
+    
+    if (j==3) {
+        std::cout<<"j has the value 3";
+    }
+    
+    if (j==3) {
+        std::cout<<"j has the value 3";
+    }
+    
+    if (j==3) {
+        std::cout<<"j has the value 3";
+    }
+    
+    j = j + 1;
+}
