+++ c2c-translator/data/translations/translation_blocks_from_python.java
@@ -1,0 +1,21 @@
+while (j < 5) {
+    System.out.println(j);
+    if (j==2) {
+        System.out.println("j has the value 2");
+        int a = 5;
+    }
+    
+    if (j==3) {
+        System.out.println("j has the value 3");
+    }
+    
+    if (j==3) {
+        System.out.println("j has the value 3");
+    }
+    
+    if (j==3) {
+        System.out.println("j has the value 3");
+    }
+    
+    j = j + 1;
+}
