+++ c2c-translator/data/translations/translation_blocks_from_python.py
@@ -1,0 +1,16 @@
+while j < 5:
+    print(j)
+    if j==2:
+        print("j has the value 2")
+        a = 5
+    
+    if j==3:
+        print("j has the value 3")
+    
+    if j==3:
+        print("j has the value 3")
+    
+    if j==3:
+        print("j has the value 3")
+    
+    j = j + 1
