+++ c2c-translator/rules.json
@@ -19,6 +19,11 @@
             "type name = value operator value;\n",
             "type name = value operator value;\n",
             "name = value operator value\n"
+        ],
+        [
+            "&&",
+            "System.out.println(@);\n",
+            "print(@)\n"
         ]
     ],
     "expression_statement": [
@@ -81,6 +86,26 @@
             "name = name operator name operator name;\n",
             "name = name operator name operator name;\n",
             "name = name operator name operator name\n"
+        ],
+        [
+            "#name operatornameoperator\n",
+            "name name name {\n",
+            "print(@)\n"
+        ],
+        [
+            "&&",
+            "System.out.println(@);\n",
+            "print(@)\n"
+        ],
+        [
+            "while",
+            "System.out.println(@);\n",
+            "if @:\n    @"
+        ],
+        [
+            "if (@) {\n    @\n}\n",
+            "System.out.println(@);\n",
+            "print(@)\n"
         ]
     ],
     "if_statement": [
@@ -93,6 +118,11 @@
             "if (@) {\n    @\n} else {\n    @\n}\n",
             "if (@) {\n    @\n} else {\n    @\n}\n",
             "if @:\n    @\nelse:\n    @\n"
+        ],
+        [
+            "while",
+            "System.out.println(@);\n",
+            "if @:\n    @"
         ]
     ],
     "while_statement": [
@@ -105,7 +135,37 @@
     "function_definition": [
         "public type name(@) {\n    @\n}\n",
         "type name(@) {\n    @\n}\n",
-        "def name(@):\n   @\n"
+        "def name(@):\n   @\n",
+        [
+            "type sum_two_numbers(type a, type b){\n",
+            "name name name {\n",
+            "name sum_two_numbers(a, b):\n"
+        ]
+    ],
+    "class_declaration": [
+        [
+            "#name operatornameoperator\n",
+            "name name name {\n",
+            "print(@)\n"
+        ],
+        [
+            "type sum_two_numbers(type a, type b){\n",
+            "name name name {\n",
+            "name sum_two_numbers(a, b):\n"
+        ]
+    ],
+    "break_statement": [
+        [
+            "    \tname::nameoperatoroperatornameoperatoroperatorname::endl;\n",
+            "          name operator value;\n",
+            "        name\n"
+        ]
+    ],
+    "for_statement": [
+        [
+            "#name operatornameoperator\n",
+            "name name name {\n",
+            "for"
+        ]
     ]
-
 }
