+++ c2c-translator/tree-keywords.txt
@@ -1,1 +1,1 @@
-local_variable_declaration,expression_statement,if_statement,declaration,while_statement
+local_variable_declaration,expression_statement,if_statement,declaration,while_statement,class_declaration,break_statement,for_statement,function_definition
